Refactored everything to bring code to base existance.
