+++ src/elements.tsx
@@ -1,27 +1,8 @@
 import { css } from "@emotion/react";
 import styled from "@emotion/styled";
 import React from "react";
-import { portfolioBackground100, portfolioActive, portfolioActiveSecondary, systemFont } from "./theme";
-
-interface ParalaxProps extends React.HTMLProps<HTMLDivElement> {
-    paralaxHeight: number;
-    viewportHeight: number;
-}
-
-export const ParalaxDiv = styled.div<ParalaxProps>`
-    position: fixed;
-    left: 0px;
-    right: 0px;
-    background-color: ${portfolioBackground100};
-
-    ${({paralaxHeight, viewportHeight}: ParalaxProps) => `
-        top: ${viewportHeight - paralaxHeight}px;
-        bottom: 0px;
-        overflow: ${viewportHeight === paralaxHeight ? 'scroll' : 'hidden'};
-    `}
-
-    transition: top 200ms;
-`;
+import { portfolioActive, portfolioActiveSecondary, systemFont, blackText, headerHeight } from "./theme";
+import { PortfolioLogo } from "./assets";
 
 export const PageContent = styled.div`
     position: absolute;
@@ -79,4 +60,100 @@ export const PrimaryCTA = styled.div`
 
 export const SecondaryCTA = styled.div`
     ${secondaryButtonCSS}
+`;
+
+export const Nav = styled.div`
+    position: absolute;
+    transition: top 50ms;
+
+    height: ${headerHeight}px;
+    width: 100%;
+    z-index: 1;
+    background-color: white;
+
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+
+    transition: top 200ms;
+`;
+
+export const NavLogo = styled(PortfolioLogo)`
+    margin: 0px 0px 0px 30px;
+`;
+
+export const NavigationLinks = styled.div`
+    margin: 0px 30px 0px 0px;
+    width: 33%;
+    height: 100%;
+
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-evenly;
+
+    @media (max-width: 780px) {
+        width: 50%;
+    }
+
+    @media (max-width: 545px) {
+        width: 75%;
+    }
+`;
+
+export interface SelectedAwareProps extends React.HTMLProps<HTMLElement> {
+    selected: boolean;
+}
+
+export const NavigationLink = styled.h2<SelectedAwareProps>`
+    position: relative;
+
+    font-family: ${systemFont};
+    font-weight: 500;
+    font-size: 19px;
+    line-height: 36px;
+    color: ${blackText};
+
+    :after {
+        content: '';
+        position: absolute;
+        left: 0px;
+        bottom: 0px;
+        width: 100%;
+        height: 3px;
+
+        background-color: ${blackText};
+
+        transition: transform 250ms;
+        transform: ${({selected}: SelectedAwareProps) => selected ? 'scaleX(1)' : 'scaleX(0)'};
+    }
+
+    ${({selected}: SelectedAwareProps) => selected ? `
+        opacity: 1;
+        cursor: default;
+    ` : `
+        opacity: 0.5;
+        cursor: pointer;
+
+        transition: opacity 250ms;
+
+        :hover {
+            opacity: 1;
+        }
+    `}
+
+    @media (max-width: 385px) {
+        font-size: 16px;
+    }
+`;
+
+export const ContentWrapper = styled.div`
+    width: 100%;
+    height: calc(100svh - ${headerHeight}px);
+    background-color: white;
+    color: #333333;
+    border-color: #333333;
+
+    padding-top: ${headerHeight}px;
 `;
