+++ src/pages/home/index.tsx
@@ -1,0 +1,11 @@
+import { HeroContent } from "./hero-content";
+import { Portfolio } from "./portfolio";
+
+export function Home() {
+    return (
+        <>
+            <HeroContent />
+            <Portfolio />
+        </>
+    );
+}
