+++ src/pages/home/portfolio/elements.ts
@@ -1,94 +1,8 @@
 import styled from "@emotion/styled";
 import React, { HTMLProps } from "react";
-import { externalLinkUrl } from "../../assets/icons";
-import { PrimaryCTA, SecondaryCTA } from "../../elements";
-import { systemFont, portfolioActive, portfolioBackground200 } from "../../theme";
-
-/***** index.tsx Stying *****/
-
-interface ActiveAwareProps extends React.HTMLProps<HTMLHeadingElement> {
-    active: boolean;
-}
-
-interface PageIndexAwareProps extends React.HTMLProps<HTMLDivElement> {
-    pageIndex: number;
-    bodyHeight: number | undefined;
-}
-
-export const PortfolioContainer = styled.div`
-    display: flex;
-    width: 100%;
-
-    flex-direction: column;
-`;
-
-export const PortfolioNav = styled.div`
-    display: flex;
-    width: 100%;
-    height: 90px;
-    background-color: ${portfolioBackground200};
-
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-evenly;
-
-    position: sticky;
-    top: 0;
-    z-index: 1;
-`;
-
-export const NavItem = styled.h2<ActiveAwareProps>`
-    font-family: ${systemFont};
-    font-weight: 600;
-    font-size: 19px;
-    line-height: 36px;
-    text-transform: capitalize;
-
-    color: ${portfolioActive};
-    transition: filter 250ms;
-    :hover {
-        filter: grayscale(0);
-    }
-
-    @media (max-width: 550px) {
-        font-size: 14px;
-    }
-
-    ${({active}: ActiveAwareProps) => active ? `
-        cursor: default;
-        filter: grayscale(0);
-    ` : `
-        cursor: pointer;
-        filter: grayscale(1);
-    `}
-
-    :after {
-        content: '';
-        position: absolute;
-        left: 0px;
-        bottom: 0px;
-        width: 100%;
-        height: 3px;
-
-        background-color: ${portfolioActive};
-
-        transition: transform 250ms;
-        transform: ${({active}: ActiveAwareProps) => active ? 'scaleX(1)' : 'scaleX(0)'};
-    }
-`;
-
-export const PortfolioBody = styled.div<PageIndexAwareProps>`
-    position: relative;
-    width: 100%;
-
-    height: ${({bodyHeight}: PageIndexAwareProps) => bodyHeight ?? 0}px;
-
-    transition: transform 400ms;
-    transform: rotateY(${({pageIndex}: PageIndexAwareProps) => (pageIndex * -90) + .001}deg);
-    transform-style: preserve-3d;
-`;
-
-/***** project.tsx Stying *****/
+import { externalLinkUrl } from "../../../assets/icons";
+import { PrimaryCTA, SecondaryCTA } from "../../../elements";
+import { portfolioActive, portfolioBackground200 } from "../../../theme";
 
 interface IndexAwareImg extends React.HTMLProps<HTMLImageElement> {
     index: number;
