+++ src/pages/home/portfolio/the-lab/index.tsx
@@ -10,32 +10,22 @@ import {
     ProjectIFrame,
     IFrameGithubLink
 } from './elements';
-import {LabBanner} from '../../../assets';
+import {LabBanner} from '../../../../assets';
 
 import {
     QuoteContainer,
     Quote,
-    QuoteCTAsContainer,
-    ProjectPrimaryCTA,
-    ProjectSecondaryCTA,
 } from '../elements';
-
-import { IAppSection } from '../../../dec';
 
 export interface TheLabProps extends React.HTMLProps<HTMLDivElement> {
     active: boolean;
-    setTopLevelPage: (newSection: IAppSection) => void;
 }
 
-export function TheLab({active, setTopLevelPage}: TheLabProps) {
+export function TheLab({active}: TheLabProps) {
     return (
         <LabContainer id="lab-projects" active={active}>
             <QuoteContainer backgroundImage={LabBanner} backgroundPositionY={'100%'} overlayOpacityHex={'80'}>
                 <Quote>Interested in someone who&lsquo;s always learning?</Quote>
-                <QuoteCTAsContainer>
-                    <ProjectPrimaryCTA onClick={() => setTopLevelPage('contact')}>Connect with me</ProjectPrimaryCTA>
-                    <ProjectSecondaryCTA onClick={() => setTopLevelPage('about')}>Read about me</ProjectSecondaryCTA>
-                </QuoteCTAsContainer>
             </QuoteContainer>
             <LabTitle>my experiments</LabTitle>
             <LabSubtitle>Projects which are either still in development, just for fun, or for personal use only</LabSubtitle>
