+++ src/pages/home/portfolio/web-projects/elements.ts
@@ -1,0 +1,14 @@
+import styled from "@emotion/styled";
+import React from "react";
+import { portfolioBackground100 } from "../../../../theme";
+
+interface ActiveAwareDiv extends React.HTMLProps<HTMLDivElement> {
+    active: boolean;
+}
+
+export const WebProjectsContainer = styled.div<ActiveAwareDiv>`
+    display: flex;
+    flex-direction: column;
+
+    background-color: ${portfolioBackground100};
+`;
