+++ src/pages/home/portfolio/web-projects/index.tsx
@@ -1,6 +1,6 @@
-import { projects } from "../../../app-data/projects";
+import { projects } from "../../../../app-data/projects";
 import { Project } from "../project";
-import {ProjectsBanner} from '../../../assets';
+import {ProjectsBanner} from '../../../../assets';
 
 import {
     WebProjectsContainer,
@@ -9,29 +9,19 @@ import {
 import {
     QuoteContainer,
     Quote,
-    QuoteCTAsContainer,
     ProjectsList,
     ProjectsTitle,
-    ProjectPrimaryCTA,
-    ProjectSecondaryCTA
 } from '../elements';
-
-import { IAppSection } from "../../../dec";
 
 export interface WebProjectProps extends React.HTMLProps<HTMLDivElement> {
     active: boolean;
-    setTopLevelPage: (newSection: IAppSection) => void;
 }
 
-export function WebProjects({active, setTopLevelPage}: WebProjectProps) {
+export function WebProjects({active}: WebProjectProps) {
     return (
         <WebProjectsContainer id="web-projects" active={active}>
             <QuoteContainer backgroundImage={ProjectsBanner} backgroundPositionY={'45%'}>
                 <Quote>Want someone to help create beautiful &#38; accessible web apps?</Quote>
-                <QuoteCTAsContainer>
-                    <ProjectPrimaryCTA onClick={() => setTopLevelPage('contact')}>Get in touch</ProjectPrimaryCTA>
-                    <ProjectSecondaryCTA onClick={() => setTopLevelPage('about')}>Learn about me</ProjectSecondaryCTA>
-                </QuoteCTAsContainer>
             </QuoteContainer>
             <ProjectsTitle>My Web Apps</ProjectsTitle>
             <ProjectsList>
